+++ cirq-core/cirq/experiments/__init__.py
@@ -71,16 +71,9 @@ from cirq.experiments.n_qubit_tomography import (
     StateTomographyExperiment,
 )
 
-from cirq.experiments.parallel_readout_calibration import (
-    estimate_parallel_readout_errors,
-)
-
-from cirq.experiments.readout_experiment_result import (
-    ReadoutExperimentResult,
-)
-
 from cirq.experiments.single_qubit_readout_calibration import (
     estimate_single_qubit_readout_errors,
+    estimate_parallel_readout_errors,
     SingleQubitReadoutCalibrationResult,
 )
