+++ cirq-core/cirq/experiments/single_qubit_readout_calibration_test.py
@@ -19,6 +19,16 @@ import numpy as np
 import cirq
 
 
+def test_single_qubit_readout_result_repr():
+   result = cirq.experiments.SingleQubitReadoutCalibrationResult(
+              zero_state_errors={cirq.LineQubit(0): 0.1},
+              one_state_errors={cirq.LineQubit(0): 0.2},
+              repetitions=1000,
+              timestamp=0.3,
+               )
+   cirq.testing.assert_equivalent_repr(result)
+
+
 class NoisySingleQubitReadoutSampler(cirq.Sampler):
     def __init__(self, p0: float, p1: float, seed: 'cirq.RANDOM_STATE_OR_SEED_LIKE' = None):
         """Sampler that flips some bits upon readout.
