+++ cirq-core/cirq/json_resolver_cache.py
@@ -118,13 +118,12 @@ def _class_resolver_dictionary() -> Dict[str, ObjectFactory]:
         'PhasedXPowGate': cirq.PhasedXPowGate,
         'PhasedXZGate': cirq.PhasedXZGate,
         'RandomGateChannel': cirq.RandomGateChannel,
-        'ReadoutExperimentResult': cirq.experiments.ReadoutExperimentResult,
         'QuantumFourierTransformGate': cirq.QuantumFourierTransformGate,
         'RepetitionsStoppingCriteria': cirq.work.RepetitionsStoppingCriteria,
         'ResetChannel': cirq.ResetChannel,
         'SingleQubitMatrixGate': single_qubit_matrix_gate,
         'SingleQubitPauliStringGateOperation': cirq.SingleQubitPauliStringGateOperation,
-        'SingleQubitReadoutCalibrationResult': cirq.experiments.ReadoutExperimentResult,
+        'SingleQubitReadoutCalibrationResult': cirq.experiments.SingleQubitReadoutCalibrationResult,
         'StabilizerStateChForm': cirq.StabilizerStateChForm,
         'SwapPowGate': cirq.SwapPowGate,
         'SymmetricalQidPair': cirq.SymmetricalQidPair,
